+++ adapters/b.pl
@@ -4,7 +4,7 @@
 %  Copyright (c) 1998-2013 Paulo Moura <pmoura@logtalk.org>
 %
 %  Adapter file for B-Prolog 7.8 and later versions
-%  Last updated on June 1, 2013
+%  Last updated on August 4, 2013
 %
 %  This program is free software: you can redistribute it and/or modify
 %  it under the terms of the GNU General Public License as published by
@@ -351,14 +351,12 @@
 	load(File).
 
 
-% '$lgt_compare_file_modification_times'(?atom, +atom, +atom)
+% '$lgt_file_modification_time'(+atom, -nonvar)
 %
-% compare file modification times
+% gets a file modification time, assumed to be an opaque term but comparable
 
-'$lgt_compare_file_modification_times'(Result, File1, File2) :-
-	file_property(File1, modification_time(Time1)),
-	file_property(File2, modification_time(Time2)),
-	compare(Result, Time1, Time2).
+'$lgt_file_modification_time'(File, Time) :-
+	file_property(File, modification_time(Time)).
 
 
 % '$lgt_environment_variable'(?atom, ?atom)