+++ adapters/cx.pl
@@ -4,7 +4,7 @@
 %  Copyright (c) 1998-2013 Paulo Moura <pmoura@logtalk.org>
 %
 %  Adapter file for CxProlog 0.97.6 or a later version
-%  Last updated on June 1, 2013
+%  Last updated on August 4, 2013
 %
 %  This program is free software: you can redistribute it and/or modify
 %  it under the terms of the GNU General Public License as published by
@@ -366,17 +366,12 @@ setup_call_cleanup(Setup, Call, Cleanup) :-
 	consult(File).
 
 
-% '$lgt_compare_file_modification_times'(?atom, +atom, +atom)
+% '$lgt_file_modification_time'(+atom, -nonvar)
 %
-% compare file modification times
-%
-% should fail if file modification times cannot be retrived 
-% or if one of the files does not exist
+% gets a file modification time, assumed to be an opaque term but comparable
 
-'$lgt_compare_file_modification_times'(Result, File1, File2) :-
-	fs_property(File1, time, [_, Time1]),
-	fs_property(File2, time, [_, Time2]),
-	compare(Result, Time1, Time2).
+'$lgt_file_modification_time'(File, Time) :-
+	fs_property(File, time, [_, Time]).
 
 
 % '$lgt_environment_variable'(?atom, ?atom)