+++ adapters/swi.pl
@@ -4,7 +4,7 @@
 %  Copyright (c) 1998-2013 Paulo Moura <pmoura@logtalk.org>
 %
 %  Adapter file for SWI Prolog 6.0.0 and later versions
-%  Last updated on June 1, 2013
+%  Last updated on August 4, 2013
 %
 %  This program is free software: you can redistribute it and/or modify
 %  it under the terms of the GNU General Public License as published by
@@ -385,14 +385,12 @@
 	load_files(Path, [derived_from(Source)| Options]).
 
 
-% '$lgt_compare_file_modification_times'(?atom, +atom, +atom)
+% '$lgt_file_modification_time'(+atom, -nonvar)
 %
-% compare file modification times
+% gets a file modification time, assumed to be an opaque term but comparable
 
-'$lgt_compare_file_modification_times'(Result, File1, File2) :-
-	time_file(File1, Time1),
-	time_file(File2, Time2),
-	compare(Result, Time1, Time2).
+'$lgt_file_modification_time'(File, Time) :-
+	time_file(File, Time).
 
 
 % '$lgt_environment_variable'(?atom, ?atom)