+++ adapters/unsupported/minerva.pl
@@ -4,7 +4,7 @@
 %  Copyright (c) 1998-2013 Paulo Moura <pmoura@logtalk.org>
 %
 %  Adapter file for Minerva 2.4
-%  Last updated on June 1, 2013
+%  Last updated on August 4, 2013
 %
 %  This program is free software: you can redistribute it and/or modify
 %  it under the terms of the GNU General Public License as published by
@@ -302,16 +302,12 @@ call(F, A1, A2, A3, A4, A5, A6) :-
 	load(File).
 
 
-% '$lgt_compare_file_modification_times'(?atom, +atom, +atom)
+% '$lgt_file_modification_time'(+atom, -nonvar)
 %
-% compare file modification times
+% gets a file modification time, assumed to be an opaque term but comparable
 
-'$lgt_compare_file_modification_times'(Result, Path1, Path2) :-
-	fileCreate(Path1, File1),
-	fileLastModified(File1, Time1),
-	fileCreate(Path2, File2),
-	fileLastModified(File2, Time2),
-	compare(Result, Time1, Time2).
+'$lgt_file_modification_time'(File, Time) :-
+	fileLastModified(File, Time).
 
 
 % '$lgt_environment_variable'(?atom, ?atom)